@@ -5,6 +5,8 @@
  * @package notification/signature
  */
 
+declare(strict_types=1);
+
 /**
  * Load the main plugin file.
  */
@@ -13,6 +15,10 @@ require_once __DIR__ . '/notification-signature.php';
 /**
  * Initialize early.
  */
-add_action( 'notification/init', function() {
-	NotificationSignature::init( __FILE__ )->init();
-}, 1 );
+add_action(
+	'notification/init',
+	static function () {
+		NotificationSignature::init(__FILE__)->init();
+	},
+	1
+);
