@@ -5,15 +5,17 @@
  * @package notification/signature
  */
 
+declare(strict_types=1);
+
 namespace BracketSpace\Notification\Signature\Requirements;
 
-use BracketSpace\Notification\Signature\Vendor\Micropackage\Requirements;
+use BracketSpace\Notification\Signature\Dependencies\Micropackage\Requirements;
 
 /**
  * BasePlugin checker
  */
-class BasePlugin extends Requirements\Abstracts\Checker {
-
+class BasePlugin extends Requirements\Abstracts\Checker
+{
 	/**
 	 * Checker name
 	 *
@@ -27,21 +29,26 @@ class BasePlugin extends Requirements\Abstracts\Checker {
 	 * @param  string $value Requirement.
 	 * @return void
 	 */
-	public function check( $value ) {
-
-		if ( ! class_exists( 'Notification' ) ) {
-			$this->add_error( 'The Notification plugin is required to be active.' );
+	public function check( $value )
+	{
+		if (! class_exists('Notification')) {
+			$this->add_error('The Notification plugin is required to be active.');
 		}
 
-		if ( version_compare( \Notification::version(), $value, '<' ) ) {
-			$this->add_error( sprintf(
+		if (! version_compare(\Notification::version(), $value, '<')) {
+			return;
+		}
+
+		$this->add_error(
+			sprintf(
 				// Translators: 1. Required Notification version, 2. Used Notification version.
-				__( 'Minimum required version of Notification plugin is %1$s. Your version is %2$s', 'notification-signature' ),
+				__(
+					'Minimum required version of Notification plugin is %1$s. Your version is %2$s',
+					'notification-signature'
+				),
 				$value,
 				\Notification::version()
-			) );
-		}
-
+			)
+		);
 	}
-
 }
