@@ -1,10 +1,18 @@
 <?php
 /**
+ * Plugin uninstall script
+ *
+ * @package notification/signature
+ */
+
+declare(strict_types=1);
+
+/**
  * Plugin uninstallation file
  *
  * @package notification/signature
  */
 
-if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+if (! defined('WP_UNINSTALL_PLUGIN')) {
 	die;
 }
